# Improve performance of DAG-CBOR encoding