## Before:

```txt
Hello World Encode:
===================
libipld  : 562 ns

Realistic Encode Tests:
=======================
canada.json.dagcbor            libipld  : 37.99 ms (26.52 MB/s)
citm_catalog.json.dagcbor      libipld  : 18.12 ms (18.02 MB/s)
twitter.json.dagcbor           libipld  : 8.11 ms (47.37 MB/s)
```

## After:

```txt
Hello World Encode:
===================
libipld  : 248 ns

Realistic Encode Tests:
=======================
canada.json.dagcbor            libipld  : 1.84 ms (546.72 MB/s)
citm_catalog.json.dagcbor      libipld  : 3.85 ms (84.78 MB/s)
twitter.json.dagcbor           libipld  : 1.94 ms (197.72 MB/s)
```

- canada.json.dagcbor **x20.6**
- citm_catalog.json.dagcbor  **x4.7**
- twitter.json.dagcbor **x4.1**